chore/routing_model: extract out flows and states
