+++ routing_model/src/utilities.rs
@@ -269,4 +269,3 @@ impl LocalEvent {
         Event::LocalEvent(*self)
     }
 }
-
